requirements.txt
Flask==3.0.3
google-cloud-bigquery==3.25.0
google-cloud-secret-manager==2.20.2
ibm-db==3.2.0
pandas==2.0.3
numpy==1.24.3
db-dtypes==1.1.1
gunicorn==21.2.0